feat: Add weighted average effective deposit calculation
This commit introduces a default weighted average effective deposit calculation. It means that it now calculates the average deposit based on the weight (hence, the importance) of each deposit in proportion to its time. As part of this process, tests and cut-off mechanisms are also put into place to ensure accuracy and reliability.